combine index.model.js and user.model.js && modify userSchema
